Fim de refatoração, alterando de Double para BigDecimal

+++ banco/src/modelo/pagamento/Boleto.java
@@ -2,19 +2,21 @@ package modelo.pagamento;
 
 import modelo.Pessoa;
 
+import java.math.BigDecimal;
+
 public class Boleto implements DocumentoPagavel {
 
     private Pessoa beneficiario;
-    private double valor;
+    private BigDecimal valor;
     private boolean estaPago;
 
-    public Boleto(Pessoa beneficiario, double valor) {
+    public Boleto(Pessoa beneficiario, BigDecimal valor) {
         this.beneficiario = beneficiario;
         this.valor = valor;
     }
 
     @Override
-    public double getValorTotal() {
+    public BigDecimal getValorTotal() {
         return valor;
     }
